SONAR-24514 Merge graph loading into persist step.
