+++ server/sonar-ce-task-projectanalysis/src/main/java/org/sonar/ce/task/projectanalysis/measure/PostMeasuresComputationCheck.java
@@ -21,6 +21,7 @@ package org.sonar.ce.task.projectanalysis.measure;
 
 import org.sonar.api.ExtensionPoint;
 import org.sonar.api.ce.ComputeEngineSide;
+import org.sonar.ce.common.scanner.ScannerReportReader;
 import org.sonar.ce.task.projectanalysis.analysis.AnalysisMetadataHolder;
 import org.sonar.ce.task.projectanalysis.analysis.Branch;
 import org.sonar.ce.task.projectanalysis.component.Component;
@@ -53,6 +54,7 @@ public interface PostMeasuresComputationCheck {
 
     Branch getBranch();
 
+    ScannerReportReader getReportReader();
     /**
      * Return the ncloc computed for the current analysis
      */
