+++ server/sonar-ce-task-projectanalysis/src/main/java/org/sonar/ce/task/projectanalysis/measure/PreMeasuresComputationChecksStep.java
@@ -19,17 +19,16 @@
  */
 package org.sonar.ce.task.projectanalysis.measure;
 
-import static org.sonar.ce.task.projectanalysis.measure.PreMeasuresComputationCheck.PreMeasuresComputationCheckException;
-
 import org.sonar.api.config.Configuration;
 import org.sonar.api.utils.System2;
-import org.sonar.ce.common.scanner.ScannerReportReader;
 import org.sonar.ce.task.log.CeTaskMessages;
 import org.sonar.ce.task.projectanalysis.analysis.AnalysisMetadataHolder;
 import org.sonar.ce.task.projectanalysis.analysis.Branch;
 import org.sonar.ce.task.projectanalysis.component.ConfigurationRepository;
 import org.sonar.ce.task.step.ComputationStep;
 import org.sonar.db.dismissmessage.MessageType;
+
+import static org.sonar.ce.task.projectanalysis.measure.PreMeasuresComputationCheck.PreMeasuresComputationCheckException;
 
 /**
  * Execute {@link PreMeasuresComputationCheck} instances in no specific order.
@@ -42,17 +41,15 @@ public class PreMeasuresComputationChecksStep implements ComputationStep {
   private final PreMeasuresComputationCheck[] extensions;
   private final CeTaskMessages ceTaskMessages;
   private final ConfigurationRepository configurationRepository;
-  private final ScannerReportReader reportReader;
 
 
   public PreMeasuresComputationChecksStep(
     AnalysisMetadataHolder analysisMetadataHolder, CeTaskMessages ceTaskMessages,
-    ConfigurationRepository configurationRepository, ScannerReportReader reportReader, PreMeasuresComputationCheck... extensions) {
+    ConfigurationRepository configurationRepository, PreMeasuresComputationCheck... extensions) {
     this.analysisMetadataHolder = analysisMetadataHolder;
     this.ceTaskMessages = ceTaskMessages;
     this.configurationRepository = configurationRepository;
     this.extensions = extensions;
-    this.reportReader = reportReader;
   }
 
   @Override
@@ -88,11 +85,5 @@ public class PreMeasuresComputationChecksStep implements ComputationStep {
     public Configuration getConfiguration() {
       return configurationRepository.getConfiguration();
     }
-
-    @Override
-    public ScannerReportReader getReportReader() {
-      return reportReader;
-    }
-
   }
 }
